Create README.md
If you have downloaded Driver-Backup.bat, you can use it to back up all the drivers you have at once.

Open the Driver-Backup.bat file just by clicking on it and all drivers will be backed up.

A folder named DriverBackup will automatically be created on your desktop.

An Install-all-drivers.bat file will also automatically be placed in the DriverBackup folder. Open the Install-all-drivers.bat file just by clicking on it  and all drivers will be installed.

It is useful, for example, if you have made a backup of the drivers of your computer, because those are missing from the execution files.
This way, you don’t need to install each device separately through Device Manager.

This saves a lot of time!